backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
aiofiles==23.1.0
psycopg2-binary==2.9.7
